Rename JobLauncherWizard to JobLauncher
