+++ src/tapis-ui/components/jobs/JobLauncher/index.tsx
@@ -1,7 +1,7 @@
 import { TapisSystem } from '@tapis/tapis-typescript-systems';
 import * as Apps from '@tapis/tapis-typescript-apps';
 
-export { default } from './JobLauncherWizard';
+export { default } from './JobLauncher';
 
 export type JobStepProps = {
   app?: Apps.TapisApp;
